Fixed call to stepsUpdated when updating steps for the current day
- Changed steps from Int16 to Int32 because Int16 is relatively small (-32,767 to 32,767)

+++ mySteps/mySteps/Models/StepsInMonth+CoreDataClass.swift
@@ -11,6 +11,6 @@ import CoreData
 
 @objc(StepsInMonthMO)
 public class StepsInMonthMO: NSManagedObject {
-    @NSManaged public var month: Int16
-    @NSManaged public var year: Int16
+    @NSManaged public var month: Int32
+    @NSManaged public var year: Int32
 }
